Make icGrep build with system/custom LLVM
Many distributions configure llvm with the default all targets, instead of x86
only in our README. The "InitializeAllTargets" API of LLVM used in
"generate_predefined_ucd_functions.cpp" requires to link against all the
configured targets. Thus, a new linking instruction is added to CMakeLists,
which is a no-op if we follow the building instructions in README, but will
fix the build when system/custom LLVM is used.



git-svn-id: http://parabix.costar.sfu.ca/svn/icGREP/icgrep-devel@4784 a82c6ad6-7e1c-4be0-8e68-26c050b43eaf